[aps-client] Improve API of PushApsClient.
This adds the following properties that will be used to initialize TLS.

  PushApsClient:ssl-cert-file
  PushApsClient:ssl-key-file
  PushApsClient:tls-certificate

ssl-cert-file and ssl-key-file may be provided, or tls-certificate may be
provided. One of those two options must be set.

Also, we add the "device-removed" signal. This signal will eventually be
fired when a device has been removed from the service.