[aps-message] Allow caching of JSON message.
This stores the json encoded string within the message so that it may
be cached for future calls to this function (which may happen if sending
the same message to a group of people). Additionally, it allows the caller
to not need to worry about the memory management.

The JSON string is invalidated upon changing other properties or adding
additional key/value pairs.