[aps-client] Emit "device-removed" upon INVALID_TOKEN response from gateway.
Since we are using the "enhanced" wire protocol for APS, we can get a
notification back from the gateway if a token is invalid. We use this
to emit the "device-removed" signal if the error was INVALID_TOKEN.

Consumers are expected to connect to this signal and remove the device
token from their database to prevent sending further notifications to
this device.

Failure to do so can yield in throttling or failure to deliver messages
sent via APS.